These files show you how I created this interactive webpage and how it can change colors using JavaScript.
In this project, I have provided files that demonstrate how I built an interactive webpage. The webpage allows users to dynamically change its colors through JavaScript functionality. You'll be able to see how JavaScript is used to manipulate the webpage's appearance, creating an engaging and customizable experience for the user